refactor: domain의 chess 폴더를 제거
