+++ src/main/java/domain/WebUIChessApplication.java
@@ -1,4 +1,4 @@
-package domain.chess;
+package domain;
 
 import spark.ModelAndView;
 import spark.template.handlebars.HandlebarsTemplateEngine;
